intern-assistant-backend/api/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.8.2
python-dotenv==1.0.1
